# Backport: new pgn reader