- backports the new pgn read

Comparison:
```
patch:

Looking (recursively) for pgn files in ..\WLD_model\pgns\
Found 11700 .pgn(.gz) files, creating 128 chunks for processing.
Progress: 128/128
Time taken for processing: 10.583s
Wrote 27779199 scored positions to popular.epd for analysis in 14.562s

master:

❯ .\fastpopular.exe -r --dir ..\WLD_model\pgns\
Looking (recursively) for pgn files in ..\WLD_model\pgns\
Found 11700 .pgn(.gz) files, creating 128 chunks for processing.
Progress: 128/128
Time taken for processing: 18.669s
Wrote 27779199 scored positions to popular.epd for analysis in 163.578s

```

I noticed that master takes a really long time to write the `popular.epd` file to the disk.
Thus I replaced `std::endl` with `\n` and it's much faster ;)

```
master with "\n" instead of std::endl

❯ .\fastpopular.exe -r --dir ..\WLD_model\pgns\
Looking (recursively) for pgn files in ..\WLD_model\pgns\
Found 11700 .pgn(.gz) files, creating 128 chunks for processing.
Progress: 128/128
Time taken for processing: 18.698s
Wrote 27779199 scored positions to popular.epd for analysis in 12.336
```

After sort(ing) the patch and master file and then running a `diff` I saw no difference.