@vondele 2023-10-13
oops, patches crossed, similar functionality merged as 5e385cd11fdbafced22e110ab702c84c81671692