@Disservin 2023-10-13
updated, the functions are now also in chronological order. I suggest to add the clang-format from the WDL repo here too, to fix the indentations